+++ src/main/java/com/example/project/repositories/ColorRepositories.java
@@ -3,6 +3,6 @@ package com.example.project.repositories;
 import com.example.project.model.Colors;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColorRepositories extends JpaRepository<Colors, Long> {
+public interface ColorRepositories extends JpaRepository<Colors, Integer> {
 
 }
